+++ .travis.yml
@@ -1,6 +1,6 @@
 language: node_js
 node_js:
-- '8.1'
+- '9'
 install:
 - npm install
 script:
