+++ package.json
@@ -21,30 +21,31 @@
   "homepage": "https://github.com/nzzdev/Q-server#readme",
   "dependencies": {
     "ajv": "^5.3.0",
-    "boom": "^4.2.0",
-    "catbox-memory": "^2.0.4",
-    "chai": "^3.5.0",
+    "boom": "^7.1.1",
+    "catbox-memory": "^3.1.1",
+    "chai": "^4.1.2",
     "clone": "^2.1.1",
     "confidence": "^3.0.2",
-    "deepmerge": "^1.5.0",
+    "deepmerge": "^2.0.1",
     "good": "^7.1.0",
     "good-console": "^6.4.0",
     "good-squeeze": "^5.0.2",
-    "h2o2": "^5.4.0",
-    "hapi": "^16.4.3",
+    "h2o2": "^6.0.1",
+    "hapi": "^17.0.1",
     "hapi-auth-couchdb-cookie": "^3.1.0",
     "hapi-etags": "^0.1.0",
-    "hapi-swagger": "^7.5.0",
-    "hoek": "^4.1.0",
-    "inert": "^4.0.3",
-    "joi": "^10.0.6",
-    "json-schema-ref-parser": "^3.1.2",
-    "mocha": "^3.2.0",
+    "hapi-swagger": "^8.0.0",
+    "hoek": "^5.0.2",
+    "inert": "^5.0.1",
+    "joi": "^13.0.1",
+    "json-schema-ref-parser": "^4.0.3",
+    "mocha": "^4.0.1",
     "nano": "^6.2.0",
+    "nano-promises": "^1.2.0",
     "node-fetch": "^1.6.3",
     "puppeteer": "^0.12.0",
-    "vision": "^4.1.1",
+    "vision": "^5.1.0",
     "white-out": "^2.0.0",
-    "wreck": "^13.0.3"
+    "wreck": "^14.0.2"
   }
 }
