+++ plugins/core/editor/routes/locales.js
@@ -3,7 +3,7 @@ const Joi = require('joi');
 module.exports = {
   path: '/editor/locales/{lng}/translation.json',
   method: 'GET',
-  config: {
+  options: {
     description: 'Returns translations for given language',
     tags: ['api', 'editor', 'non-critical'],
     validate: {
@@ -12,7 +12,7 @@ module.exports = {
       }
     }
   },
-  handler: (request, reply) => {
+  handler: (request, h) => {
     const tools = request.server.settings.app.tools.get('');
 
     // compute a translation.json file for use by i18next for the given language
@@ -26,6 +26,6 @@ module.exports = {
       translations[toolName] = tool.editor.label_locales[request.params.lng];
     }
 
-    reply(translations);
+    return translations;
   }
 }
