--- server.js
@@ -1,7 +1,0 @@
-module.exports.getServer = function() {
-  return server;
-};
-
-module.exports.setServer = function(newServer) {
-  server = newServer;
-};
